Packaging with RPM
This commit adds the capability to easily build an RPM for this tool.  The process to do it is to run ./build-rpm.sh.  Internally, this script prepares a tarball for the spec file, copies the files which are tracked by git into an rpmbuild SOURCES directory, builds an SRPM then uses the Fedora 25 Mock configuration to build the actual RPM.  This RPM should then be used on other systems.  I need to verify this, but we require a C++11 compiler to build the node modules, which is not currently possible.  I need to verify that the resulting RPM from a Fedora 25 build host actually runs on a CentOS 6 machine.  If it doesn't work, we could probably figure out which libraries are missing and copy those libraries into this package.  I doubt we'll need that though.  Maybe someone with a real centos 6 machine can double check this